Replace the manual skewing code with the BALANCED sharding strategy (#28)
The BALANCED sharding strategy ensures balance among many streams on the
server-side, obviating the need for the client to mimic such behavior. Note that
I'm using an unknown field to force the BALANCED sharding strategy. We are in
the process of releasing new Java libraries. Once the new libraries are out, I
will send another pull request to use the generated code for setting the
sharding strategy.