Adding read from query capability (#305)
Adding the option to run an ad-hoc query on BigQuery and then read the result directly into a dataframe. The syntax is in the following manner:
```
df = df = spark.read.format("bigquery")
      .option("viewsEnabled", true)
      .option("materializationDataset", <dataset>)
      .load("SELECT ...")
```
or
```
df = df = spark.read.format("bigquery")
      .option("viewsEnabled", true)
      .option("materializationDataset", <dataset>)
      .option("query", "SELECT ...")
      .load()
```
The queries are materialized to a temporary table and are re-used if the same query is used. For example
`df.select("col1").where("condition")` should use the temporary table and not re-run the query.

This also gives the option to efficiently run complex queries using joins on BigQuery specific functions from spark.