Fix next() to be interrupt safe. (#402)
for some reason I did not think Spark would interrupted the thread calling the next() method, I had trouble tracking this down but thinking more thoroughly about it I think spark could and probably does do this.  In this case if the thread was interrupted while trying to get the next batch it is possible for a memory leak to occur.

Unfortunately, the unit test change didn't replicate the flakiness, but I still think these changes make the code safer.