feat: add support for writing JSON columns through Avro indirect write
Closes #637

BigQuery supports writing JSON columns through the Avro string type as long as there is a `sqlType` type annotation of `JSON` in the schema: https://cloud.google.com/bigquery/docs/loading-data-cloud-storage-avro#extract_json_data_from_avro_data. This change reads the metadata of `StructField`s with a string type when converting from Spark SQL schema to Avro and if it contains the `JSON` for the `sqlType` key then it will include the type annotation.