Fixing APPEND setCategory issue and remove redundant load operation (#1400)
This change ensures that SaveMode.Append no longer attempts to modify the destination table's metadata, resolving the permission issue and aligning the connector's behavior with user expectations, fixing the issue #1379.

Also, this PR fixed a minor bug that has been just find out about that the call to updateMetadataIfNeeded() in BigQueryIndirectDataSourceWriterContext are not within a conditional check, which causes redundant load. The metadata update will now only be triggered when the writeDisposition is WRITE_TRUNCATE (i.e., when SaveMode is Overwrite).